+++ src/casts.rs
@@ -2,19 +2,19 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use crate::{inline::InlineString, SmartStringMode};
+use crate::{boxed::BoxedString, inline::InlineString};
 
-pub(crate) enum StringCast<'a, Mode: SmartStringMode> {
-    Boxed(&'a Mode::BoxedString),
+pub(crate) enum StringCast<'a> {
+    Boxed(&'a BoxedString),
     Inline(&'a InlineString),
 }
 
-pub(crate) enum StringCastMut<'a, Mode: SmartStringMode> {
-    Boxed(&'a mut Mode::BoxedString),
+pub(crate) enum StringCastMut<'a> {
+    Boxed(&'a mut BoxedString),
     Inline(&'a mut InlineString),
 }
 
-pub(crate) enum StringCastInto<Mode: SmartStringMode> {
-    Boxed(Mode::BoxedString),
+pub(crate) enum StringCastInto {
+    Boxed(BoxedString),
     Inline(InlineString),
 }